Fix indentation, remove teacher module, answer question about args.fetch
