+++ student.rb
@@ -1,5 +1,3 @@
-require_relative 'person'
-
 class Student < Person
 
   def initialize(options = {})
